use of latitude and longitude changed to use of class LatLng

+++ app/src/main/java/com/example/weatherapp/data/mappers/geocode/AddressToGeocodingResultMapper.kt
@@ -19,7 +19,7 @@ class AddressToGeocodingResultMapperImpl @Inject constructor() : AddressToGeocod
      */
     override fun map(valueToMap: Address): GeocodingResult = with(valueToMap) {
         GeocodingResult(
-            LatLng(latitude, longitude),
+            LatLng(latitude.toFloat(), longitude.toFloat()),
             featureName,
             countryName
         )
